requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0

google-generativeai==0.3.0
flask-cors==4.0.0
openai>=1.3.0
httpx>=0.25.0
